Renamed and moved resource view components
